requirements.txt
asgiref==3.6.0
Django==4.2.1
django-jazzmin==2.6.0
freetype-py==2.3.0
Pillow==9.5.0
pycairo==1.23.0
pypng==0.20220715.0
qrcode==7.4.2
reportlab==4.0.0
rlPyCairo==0.2.0
sqlparse==0.4.4
typing_extensions==4.5.0
